aaaaaaaaaaaaaa

+++ src/displayapp/screens/Stopwatch.h
@@ -25,7 +25,7 @@ namespace Pinetime {
                 
                 std::chrono::_V2::system_clock::time_point startTime = std::chrono::system_clock::now();
                 std::chrono::_V2::system_clock::time_point endTime = std::chrono::system_clock::now();
-                double elapsedTime = 0;
+                float elapsedTime = 0;
 
                 void startTimer();
                 void stopTimer();
